requirements.txt
dash==2.16.1
plotly==5.20.0
pandas==2.2.2
gunicorn==21.2.0
psycopg2-binary==2.9.9